add some components and folder structure to components, deleted all HTML from Header.jsx and index.html
