+++ src/App.jsx
@@ -1,10 +1,13 @@
-import Header from "@/components/Header.jsx";
+import Header from "@/components/header/Header.jsx";
+import Main from "@/components/main/Main.jsx";
+import Footer from "@/components/footer/Footer.jsx";
 
 const App = () => {
   return (
     <>
-      <div className="text-4xl font-bold text-success">Hallo på do!</div>
       <Header />
+      <Main />
+      <Footer />
     </>
   );
 };
