+++ src/components/footer/Footer.jsx
@@ -1,0 +1,4 @@
+const Footer = () => {
+  return;
+};
+export default Footer;
