+++ src/components/header/Logo.jsx
@@ -1,0 +1,4 @@
+const Logo = () => {
+  return <h1>Portfolio</h1>;
+};
+export default Logo;
