+++ src/components/main/Main.jsx
@@ -1,0 +1,4 @@
+const Main = () => {
+  return;
+};
+export default Main;
