+++ tailwind.config.js
@@ -1,7 +1,7 @@
 export default {
   content: ["./index.html", "./src/**/*.{js,jsx,ts,tsx}"],
   safelist: [],
-  darkMode: "selector",
+  darkMode: "class",
   theme: {}, // In TailwindCSS v4 every custom theme-style must be under @theme in index.css
   plugins: [],
 };
